# contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/videos": {
      "get": {
        "operationId": "VideosController_list",
        "parameters": [
          {
            "name": "filter",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "all",
                "processing",
                "ready",
                "failed"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "videos"
        ]
      },
      "post": {
        "operationId": "VideosController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVideoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create the durable video record before any long-running processing begins.",
        "tags": [
          "videos"
        ]
      }
    },
    "/api/videos/{id}": {
      "get": {
        "operationId": "VideosController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "videos"
        ]
      },
      "delete": {
        "operationId": "VideosController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": ""
          }
        },
        "tags": [
          "videos"
        ]
      }
    },
    "/api/videos/{id}/upload-session": {
      "post": {
        "operationId": "VideosController_uploadSession",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "summary": "Signed direct-to-storage upload information (single PUT or multipart). The file never transits the API.",
        "tags": [
          "videos"
        ]
      }
    },
    "/api/videos/{id}/upload-session/parts": {
      "post": {
        "operationId": "VideosController_signParts",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignPartsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "videos"
        ]
      }
    },
    "/api/videos/{id}/upload-complete": {
      "post": {
        "operationId": "VideosController_uploadComplete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadCompleteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Idempotent; verifies the object exists and enqueues ingestion exactly once.",
        "tags": [
          "videos"
        ]
      }
    },
    "/api/videos/{id}/process": {
      "post": {
        "operationId": "VideosController_process",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": ""
          }
        },
        "tags": [
          "videos"
        ]
      }
    },
    "/api/videos/{id}/analyze": {
      "post": {
        "operationId": "VideosController_analyze",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": ""
          }
        },
        "tags": [
          "videos"
        ]
      }
    },
    "/api/videos/{id}/candidates": {
      "get": {
        "operationId": "VideosController_candidates",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "score",
                "time"
              ]
            }
          },
          {
            "name": "minScore",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "maximum": 100,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "videos"
        ]
      }
    },
    "/api/videos/{id}/transcript": {
      "get": {
        "operationId": "VideosController_transcript",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startMs",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          },
          {
            "name": "endMs",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "videos"
        ]
      }
    },
    "/api/videos/{id}/playback": {
      "get": {
        "operationId": "VideosController_playback",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "videos"
        ]
      }
    },
...
  },
  "info": {
    "title": "ClipRover API",
    "description": "Product API for the ClipRover AI short-form clipping platform. All endpoints require a session cookie (cr_session) or Bearer token unless marked public. Errors use the envelope { error: { code, message, retryable, correlationId } }.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "cr_session"
      },
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateVideoDto": {
        "type": "object",
        "properties": {
          "sourceType": {
            "type": "string",
            "description": "\"upload\" for direct-to-storage upload, \"url\" for a supported source link.",
            "enum": [
              "upload",
              "url"
            ]
          },
          "originalFilename": {
            "type": "string",
            "maxLength": 255
          },
          "contentType": {
            "type": "string",
            "maxLength": 100
          },
          "sizeBytes": {
            "type": "number",
            "minimum": 1
          },
          "sourceUrl": {
            "type": "string",
            "maxLength": 2048
          },
          "title": {
            "type": "string",
            "maxLength": 200
          },
          "rightsConfirmed": {
            "type": "boolean",
            "description": "The user confirms they own or are authorized to process this content."
          }
        },
        "required": [
          "sourceType",
          "rightsConfirmed"
        ]
      },
      "SignPartsDto": {
        "type": "object",
        "properties": {
          "partNumbers": {
            "minItems": 1,
            "maxItems": 100,
            "type": "array",
            "items": {
              "type": "number",
              "minimum": 1
            }
          }
        },
        "required": [
          "partNumbers"
        ]
      },
      "CompletedPartDto": {
        "type": "object",
        "properties": {
          "partNumber": {
            "type": "number",
            "minimum": 1
          },
          "etag": {
            "type": "string",
            "maxLength": 200
          }
        },
        "required": [
          "partNumber",
          "etag"
        ]
      },
      "UploadCompleteDto": {
        "type": "object",
        "properties": {
          "parts": {
            "maxItems": 10000,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompletedPartDto"
            }
          },
          "observedSizeBytes": {
            "type": "number",
            "description": "Size the client observed locally (informational; storage is authoritative).",
            "minimum": 0
          }
        }
      },
...
    }
  }
}